add built tags
