+++ pkg/util/loop/loop.go
@@ -5,11 +5,6 @@
 
 package loop
 
-import (
-	"github.com/sylabs/singularity/internal/pkg/buildcfg"
-	"github.com/sylabs/singularity/pkg/util/singularityconf"
-)
-
 // Device describes a loop device
 type Device struct {
 	MaxLoopDevices int
@@ -68,20 +63,4 @@ type Info64 struct {
 	CryptName      [64]byte
 	EncryptKey     [32]byte
 	Init           [2]uint64
-}
-
-func GetMaxLoopDevices() int {
-	// if the caller has set the current config use it
-	// otherwise parse the default configuration file
-	cfg := singularityconf.GetCurrentConfig()
-	if cfg == nil {
-		var err error
-
-		configFile := buildcfg.SINGULARITY_CONF_FILE
-		cfg, err = singularityconf.Parse(configFile)
-		if err != nil {
-			return 256
-		}
-	}
-	return int(cfg.MaxLoopDevices)
 }
