+++ pkg/util/loop/loop_no_singularity.go
@@ -1,0 +1,9 @@
+// +build !singularity_engine
+
+package loop
+
+// GetMaxLoopDevices Return the maximum number of loop devices allowed
+func GetMaxLoopDevices() int {
+	// externally imported package, use the default value
+	return 256
+}
